mobasher/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
pgvector>=0.4.0

# Queue and workers
celery[redis]>=5.3.0
redis>=5.0.0
flower>=2.0.0

# AI/ML
faster-whisper>=0.10.0
torch>=2.1.0
torchvision>=0.16.0
sentence-transformers>=2.2.0
transformers>=4.35.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
ultralytics>=8.0.0
insightface>=0.7.3
onnxruntime-gpu>=1.16.0

# OCR
paddleocr>=2.7.0
easyocr>=1.7.0

# Audio/Video processing
ffmpeg-python>=0.2.0
av>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0

# Voice Activity Detection
silero-vad>=4.0.0
webrtcvad>=2.0.10

# Observability
prometheus-client>=0.19.0
structlog>=23.2.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-celery>=0.42b0

# Utilities
tenacity>=8.2.0
pyyaml>=6.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.2.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# CLI
typer[all]>=0.12.0

# Optional: for testing with real services
testcontainers[postgresql,redis]>=3.7.0
